requirements.txt
playwright
google-generativeai
python-dotenv
schedule
beautifulsoup4
google-api-python-client
